pyproject.toml
[project]
name = "debate-fastapi-reference"
version = "0.2.0"
description = "FastAPI sidecar for debaterhub-sdk — WebSocket-driven IPDA debate server"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "David Gonier" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.0.0",
    "debaterhub-sdk @ git+https://github.com/dgonier/debaterhub-sdk.git",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
]

[project.urls]
Homepage = "https://github.com/dgonier/debate-fastapi-reference"
Repository = "https://github.com/dgonier/debate-fastapi-reference"
SDK = "https://github.com/dgonier/debaterhub-sdk"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
